Changes to use of local data and handling images
Changed code to load local data first and then start any process to check the server. This will give a better response if the user is on a low-bandwidth data connection. Further, corrected code in UIImageView extension to correctly set image to nil if there is no image for the given name; this will force use of a default image. Also updated deployment target data to 12.0 across different targets.